[chord_diagram] `order` keyword argument applies to `colors` (#32)
This PR fixes an issue where, using both `names` and `colors` arguments with a user-defined `order`, the mapping between names and colors would change from the default mapping.

E.g. with

names=["A","B","C"]
colors=["red","green","blue"]

One would not get "A"->"red", "B"->"green", "C"->"blue" if the order is not [0,1,2] but [2,1,0].

With the new change, one always gets "A"->"red", "B"->"green", "C"->"blue" regardless of the order.

Co-authored-by: gph82 <guille.perez.hernandez@gmail.com>
Reviewed-on: https://codeberg.org/tfardet/mpl_chord_diagram/pulls/32
Reviewed-by: Tanguy Fardet <tfardet@noreply.codeberg.org>
Co-authored-by: gph82 <gph82@noreply.codeberg.org>
Co-committed-by: gph82 <gph82@noreply.codeberg.org>